always return a value
The previous typing improvements caught an error in logic: `fetchResource` returns `undefined` when there is an inner error object but it has a code other than 40001 or 20166. We always want this method to either resolve to a fetched resource or throw an error, so we fix this error by retrying if the error code is something else, instead of returning undefined.