429s don't count towards retries
It is trivially easy to hit the 5 max retry limit if getting a lot of 429 backoff errors. If we have a lot of resources to fetch and we're getting heavily throttled, we should just keep retrying until it works. 

There is no reason to have a max number of retries, and even if there were, 5 is *WAY* too low.

I'm doing the thing I think is more sensible and changing the logic so that we fail permanently if we hit 5 errors, but we keep retrying as long as the only error we're getting is a wrapped error with code 20166.