429s don't count towards retries
It is trivially easy to hit the 5 max retry limit when we get a lot of 429 backoff errors. If we have a lot of resources to fetch and we're getting heavily throttled, we should just keep retrying until it works. 

There is no reason to have a max number of backoff-related retries, and even if there were, 5 is *WAY* too low. So option 1 is to have a separate, much higher retry limit for backoff errors, and option 2 is to just not count backoff errors towards the single retry limit.

I'm choosing option 2 and changing the logic so that we fail permanently if we hit 5 regular errors, but we keep retrying, potentially forever, if we keep hitting backoff errors.