fix retry timer
The previous logic didn't work. If the `Retry-After` header wasn't present, it would still use `1`, because `+!null` is `1`. The fallback value of 15s would never be used.

This isn't great because having a default retry of 1s causes us to hammer the OneNote API too much.

So, fix the logic to use 15s instead.