stop retrying requests once import is cancelled
I noticed that, because I changed it so that we keep retrying on backoff errors, there is currently no logic to STOP retrying after the import is cancelled by the user. In-flight network requests just keep retrying, potentially forever. Realistically they would eventually succeed or fail, but this isn't guaranteed.

To fix: 
- add an abort controller to the OneNote importer, use it in all network requests
- if the import is cancelled, abort

NOTE: I didn't use the `ImportContext` cancelled flag because that requires polling the flag, which would potentially make us slow to react and might be error prone. Also, it's not available everywhere in the importer (it's not available during `init`, for example) so this seemed like a smaller, less error-prone change.