add retries for fetch errors
in a big enough import I will invariably see the `fetch` hard-fail with a timeout (`net::ERR_TIMED_OUT`). I don't know if that's my network connection or the graph API, but it does mean this occasionally happens and should be handled better.

Retrying in the case where `fetch` fails seems reasonable.