OneNote: Improve typing, timeouts, and retry logic when fetching from the OneNote API. (#388)
The will greatly increase the amount of time the client will wait before giving up when hitting rate limits from the OneNote API. It also makes the retry logic more solid and fixes several smaller bugs in this part of the code.

If no successful files are fetched for 90 minutes then the import is cancelled.