Notion API(WIP): Processing external link type module
Strategy for handling modules with YouTube and Twitter (X) links in Notion:

If the link appears in Notion within modules such as Video/Embed that render the content, then in Obsidian also use the external image syntax to render the content directly.

If the link is in modules like Bookmark or Link_Preview, render it as a regular Markdown external link.

This strategy preserves the user's intended presentation.
Also: Files that fail to download will be directly linked using Markdown's external link syntax.