Notion API(WIP): Process Person content
Markdown syntax is not supported in YAML properties, so for Mention or Person types only the name is displayed. In the body text, if a Mention includes an email address, it will be shown in the format `[name](mailto:xxx)`. Note: Notion integration requires permission to read users' names and email addresses.