Airtable API: Refactor Base import process
Previously: After retrieving all complete information, assemble the data in memory and then write it to a file.
Now: Assemble the data by Base and then write it, to reduce memory usage.
Note: Because in Airtable, RecordId and FieldId are only guaranteed to be unique within the same Base, we prepend a BaseId when using them.